transfer

+++ src/main/java/com/llm/transfer/repository/TransferRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.llm.transfer.model.Transfer;
 
-public interface TransferRepository extends JpaRepository<Transfer, Integer> {
+public interface TransferRepository extends JpaRepository<Transfer, Long> {
 
 }
